requirements.txt
flask
flask-restx
flask-jwt-extended
flask-sqlalchemy
python-dotenv
pymysql
waitress